Refactored HypixelRankPlusColor to HypixelColors and added documentation
The HypixelColors class can be used for more general purposes than just rank plus colors, such as pet colors, guild tag colors and rank colors.
